requirements.txt
Flask
Pillow
pillow-heif
gunicorn  